test(GiniCaptureSDK): Add CameraViewController test to the new class
- Remove unrelated test from CameraPreviewControllerTest

PP-3289
